* Replaced isSealed() checks with our own boolean instance variable in NpcConversationActions. I realized you can't seal it before throwing the continuation, but I can't throw the continuation before sealing it. I didn't catch this before because it's a very rare occurrence to have the script forcefully halted while in the middle of evaluation (and not while waiting for a response) - Rhino/JavaScript isn't THAT slow. I'm probably going to have to also synchronize the terminated variable, but I will leave it the way it is until I start get cases of it acting up.  * NpcConversationActions.askAvatar now returns the player's selection.  * Some InventorySlot changes. Unique IDs are now long. Corrected pet packet structure - footholds are now short, and the other two bytes were actually bools. Ring's partnerRingUid are now long (even though the current packet structure for rings gives it as an int).  * Renamed GameHandler to GameEnterHandler to disambiguate it's purpose somewhat (since it handles connections and map entering).  * Completely rewrote interserver communication to accommodate splitted worlds, i.e. so that all channels of one world don't have to be running on the same process (so that channels can be spread out across many servers). Modified game0.properties, game1.properties, and added game2.properties to demonstrate how to use the new features.  * Restructured classes for interserver communications. Classes only required by the Center Server are now in the argonms.center package, and only the classes required by the remote servers (e.g. abstract classes) or required by both the Center Server and remote servers (e.g. packet header values) are located in argonms.net.server. Many classes have been renamed and overall there has been a huge refactoring in the structure.

+++ trunk/argon.sql
@@ -84,7 +84,7 @@ CREATE TABLE `inventoryitems` (
   `position` smallint(5) NOT NULL,
   `itemid` int(11) NOT NULL,
   `expiredate` int(11) UNSIGNED NOT NULL,
-  `uniqueid` int(11) UNSIGNED NOT NULL,
+  `uniqueid` bigint(20) UNSIGNED NOT NULL,
   `owner` tinytext DEFAULT NULL,
   `quantity` smallint(5) NOT NULL, /* separate this into a `inventoryitems` table and rename this one `inventories`? */
   PRIMARY KEY (`inventoryitemid`),
@@ -150,7 +150,7 @@ CREATE TABLE `inventoryrings` (
   `inventoryringid` int(10) UNSIGNED NOT NULL AUTO_INCREMENT,
   `inventoryitemid` int(10) UNSIGNED NOT NULL,
   `partnerchrid` int(11) NOT NULL,
-  `partnerringid` int(11) NOT NULL,
+  `partnerringid` bigint(20) NOT NULL,
   PRIMARY KEY (`inventoryringid`),
   KEY `inventoryitemid` (`inventoryitemid`),
   CONSTRAINT `inventoryrings_ibfk_1` FOREIGN KEY (`inventoryitemid`) REFERENCES `inventoryitems` (`inventoryitemid`) ON DELETE CASCADE
